requirements.txt
requests~=2.25.1
pandas~=1.2.4
ffmpeg-python